Move the `cmp` function into the `vals` package
This moves the value comparison logic into `pkg/eval/vals` so it can be
used by packges other than `eval`. For example, the vals package `repr`
implementation as well as other Elvish modules which may want to compare
Elvish values. This is a baby step to augmenting all the code which
order values, or might want to (e.g., the `pprint` command), to also
handle ordering otherwise uncomparable types.

Related #1495
